refactor: apply re-export
