+++ src/api/auth/dto/index.ts
@@ -1,0 +1,5 @@
+export { LoginRequestDto } from './login-request.dto';
+export { LoginResponseDto } from './login-response.dto';
+export { ActivateRequestDto } from './activate-request.dto';
+export { RegisterRequestDto } from './register-request.dto';
+export { RegisterResponseDto } from './register-response.dto';
