+++ src/api/auth/guards/index.ts
@@ -1,0 +1,1 @@
+export { JwtAuthGuard } from './jwt-auth.guard';
