+++ src/api/auth/strategies/index.ts
@@ -1,0 +1,1 @@
+export { JwtStrategy } from './jwt.strategy';
