+++ src/decorators/index.ts
@@ -1,0 +1,1 @@
+export { Public } from './public.decorator';
